Add EffectsRepository
