+++ feature/effect_detail/src/main/java/com/tomaszrykala/recsandfx/feature/effect_detail/EffectDetailViewModel.kt
@@ -4,35 +4,27 @@ import android.content.Context
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import com.tomaszrykala.recsandfx.core.domain.effect.Effect
-import com.tomaszrykala.recsandfx.core.domain.effect.toParam
 import com.tomaszrykala.recsandfx.core.domain.native.NativeInterfaceWrapper
 import com.tomaszrykala.recsandfx.core.domain.native.NativeInterfaceWrapperImpl
+import com.tomaszrykala.recsandfx.core.domain.repository.EffectsRepository
+import com.tomaszrykala.recsandfx.core.domain.repository.EffectsRepositoryImpl
 import com.tomaszrykala.recsandfx.core.storage.FileStorage
 import com.tomaszrykala.recsandfx.core.storage.FileStorageImpl
 import com.tomaszrykala.recsandfx.feature.media_player.RecordingsPlayer
 import com.tomaszrykala.recsandfx.feature.media_player.RecordingsPlayerImpl
 
 class EffectDetailViewModel(
-    private val nativeInterface: NativeInterfaceWrapper = NativeInterfaceWrapperImpl(),
     private val fileStorage: FileStorage = FileStorageImpl(),
     private val recordingsPlayer: RecordingsPlayer = RecordingsPlayerImpl(),
+    private val effectsRepository: EffectsRepository = EffectsRepositoryImpl(),
+    private val nativeInterface: NativeInterfaceWrapper = NativeInterfaceWrapperImpl(),
 ) : ViewModel() {
 
     private lateinit var effect: Effect
 
-    private val allEffects: List<Effect> by lazy {
-        nativeInterface.getAllEffectsMap().map {
-            Effect(
-                id = it.value.id,
-                name = it.key,
-                category = it.value.category,
-                params = it.value.paramValues.map { pd -> pd.toParam() },
-            )
-        }
-    }
-
+    // TODO FLOW
     fun getEffect(effectName: String): Effect? {
-        return allEffects.find { it.name == effectName }?.also {
+        return effectsRepository.getAllEffects().find { it.name == effectName }?.also {
             nativeInterface.addEffect(it)
             effect = it
         }
